refactor: Remove index-handlers dependency from test package
