+++ index-handlers/src/test/java/no/sikt/nva/nvi/index/IndexDocumentHandlerTest.java
@@ -5,12 +5,12 @@ import static java.util.Collections.emptyMap;
 import static no.sikt.nva.nvi.common.utils.JsonPointers.JSON_PTR_AFFILIATIONS;
 import static no.sikt.nva.nvi.common.utils.JsonPointers.JSON_PTR_BODY;
 import static no.sikt.nva.nvi.common.utils.JsonPointers.JSON_PTR_TYPE;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.GZIP_ENDING;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.HARD_CODED_INTERMEDIATE_ORGANIZATION;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.HARD_CODED_TOP_LEVEL_ORG;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.NVI_CANDIDATES_FOLDER;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.expandApprovals;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.expandPublicationDetails;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.GZIP_ENDING;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.HARD_CODED_INTERMEDIATE_ORGANIZATION;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.HARD_CODED_TOP_LEVEL_ORG;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.NVI_CANDIDATES_FOLDER;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.expandApprovals;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.expandPublicationDetails;
 import static no.sikt.nva.nvi.test.QueueServiceTestUtils.createEvent;
 import static no.sikt.nva.nvi.test.QueueServiceTestUtils.createEventWithOneInvalidRecord;
 import static no.sikt.nva.nvi.test.TestConstants.EN_FIELD;
