+++ index-handlers/src/test/java/no/sikt/nva/nvi/index/apigateway/FetchInstitutionReportHandlerTest.java
@@ -6,6 +6,15 @@ import static com.google.common.net.MediaType.ANY_APPLICATION_TYPE;
 import static com.google.common.net.MediaType.MICROSOFT_EXCEL;
 import static com.google.common.net.MediaType.OOXML_SHEET;
 import static java.util.Objects.nonNull;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentMissingApprovals;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentMissingCreatorAffiliationPoints;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentWithLanguage;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentWithoutIssn;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentWithoutLanguage;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentWithoutOptionalPublicationChannelData;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.indexDocumentWithoutPages;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.randomCristinOrgUri;
+import static no.sikt.nva.nvi.index.IndexDocumentTestUtils.randomIndexDocumentWith;
 import static no.sikt.nva.nvi.index.apigateway.utils.AggregateResponseTestUtil.filterAggregate;
 import static no.sikt.nva.nvi.index.apigateway.utils.ExcelWorkbookUtil.extractLinesInInstitutionIdentifierColumn;
 import static no.sikt.nva.nvi.index.apigateway.utils.ExcelWorkbookUtil.fromInputStream;
@@ -39,15 +48,6 @@ import static no.sikt.nva.nvi.index.model.report.InstitutionReportHeader.PUBLICA
 import static no.sikt.nva.nvi.index.model.report.InstitutionReportHeader.PUBLISHED_YEAR;
 import static no.sikt.nva.nvi.index.model.report.InstitutionReportHeader.REPORTING_YEAR;
 import static no.sikt.nva.nvi.index.query.SearchAggregation.TOTAL_COUNT_AGGREGATION_AGG;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentMissingApprovals;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentMissingCreatorAffiliationPoints;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentWithLanguage;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentWithoutIssn;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentWithoutLanguage;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentWithoutOptionalPublicationChannelData;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.indexDocumentWithoutPages;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.randomCristinOrgUri;
-import static no.sikt.nva.nvi.test.IndexDocumentTestUtils.randomIndexDocumentWith;
 import static no.unit.nva.commons.json.JsonUtils.dtoObjectMapper;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static no.unit.nva.testutils.RandomDataGenerator.randomUri;
